LIBITD-1869. Removed "suffix" from allowed params in controller
Removed "suffix" from the allowed params in the Handle controller. This
should prevent the suffix from being added or changed via the GUI or
REST API.

Modified the form to remove the form field for suffix, displaying
a message for "new" handles and the suffix value for "edit" handles.

Added a test to verify that the controller cannot modify the suffix.

https://issues.umd.edu/browse/LIBITD-1869