LIBITD-1869. Added hard-coded prefixes and repos to Handle model
Added hard-coded prefix and repos to the Handle model.

Updated the Handle controller to not allow changes to the prefix when
doing an update. This is because the prefix and suffix are basically
tied together, and it does not make sense to change the prefix when
you cannot also change the suffix.

Changed the form to display the "Prefix" field as a drop-down for
"new" handle creation, and as a simple label when editing.

Updated the tests.

Modified the "populate_sample_data" Rake task to use the "prefixes"
and "repos" from the Handle model.

https://issues.umd.edu/browse/LIBITD-1869