LIBITD-1870. Updated application.rb configuration and tests
The change to "config.hosts" in the "config/application.rb" file was
causing the controller, integration, and system tests to fail with
a "403 Forbidden" error, because the host being used by the tests
was not recognized by the system.

Modified the "config/application.rb" file to not modify the
"config.hosts" environment variable when running the tests.

Updated the controller, integration, and system tests to "sign in" as
a known user, so that the tests would succeed.

Added an integration test for the "ping" endpoint, to provide a
integration test of an endpoint that does not require authentication.

https://issues.umd.edu/browse/LIBITD-1870