LIBITD-1870. Modified "HOST" env var dependency to log warning
Modified the "config/application.rb" and "config/initializers/devise.rb"
configuration files to not be strictly dependent on the "HOST"
environment variable being present.

There are several cases in which these configuration are run, but that
a HOST environment variable is not needed such as:

* Running the tests
* Running database migrations
* Building the Docker image

The "HOST" environment variable is still required when actually running
the server, but there is no good way to determine if Rails is running
as a server application, so just sending warning to the log.

Removed "HOST" from the "Jenkinsfile" as it is not needed.

https://issues.umd.edu/browse/LIBITD-1870