LIBITD-1904. Fix for CSRF token issue when using Turbolinks with forms
Fixed a CSRF token verification failure when submitting a form that
had been added to the page by Turbolinks, as suggested in

https://stackoverflow.com/a/64257384

It is not 100% clear why this change is necessary, as it seems
like something that Rails should already be doing, but it works for now.

Also switched the order of stylesheet_link_tag/javascript_pack_tag and
csrf_meta_tags/csp_meta_tag includes in "_umd_lib.html.erb" to better
match layout in a default Rails 6 application.

https://issues.umd.edu/browse/LIBITD-1904