LIBITD-1906. Updated "config.host" to include K8S_INTERNAL_HOST
Rails 6 has a new feature to prevent "DNS Rebinding" attacks, described
in https://github.com/rails/rails/pull/33145, by allowing specific
hostnames to be added to an allowed list, "config.host" in the
"config/application.rb" file. If the hostname of the request does not
match a name in "config.host" the request is rejected.

By default, the "config.host" allow list is not set, and allows
connections for any host.

As part of the LIBITD-1870, the "config.host" allow list was added to
use the "HOST" environment variable. Added a "K8s_INTERNAL_HOST"
environment variable, which is added to the "config.host" in the
"config/application.rb" file, if present. This enables the
Kubernetes internal host name to be set in the k8s-umd-handle
configuration.

In the "env_example" file, provided a default value of "umd-handle-app"
for the "K8S_INTERNAL_HOST", as that seems likely to be correct
(and does not matter in the local development environment).

https://issues.umd.edu/browse/LIBITD-1906