LIBITD-1917. Added "exists" endpoint for checking whether handle exists
Added a "handles/exists" endpoint that takes "repo" and "repo_id" HTTP
parameters to determine if a handle with the given values exists. HTTP
parameters are being used in case this endpoint needs to be extended in
the future to take, for example, a static URL, or other lookup field.

Updated the "docs/umd-handle-open-api-v1.yml" OpenAPI specification for
the new endpoint.

https://issues.umd.edu/browse/LIBITD-1917