LIBITD-1917. Added "info" endpoint for returning handle information
Added a "handles/info" endpoint that takes "prefix" and "suffix" HTTP
parameters to determine if a handle with the given values exists, and
returns information about the handle. This endpoint is very similar
to the "exists" endpoint, just taking "prefix"/"suffix" instead of
"repo"/"repo_id". This endpoint is currently only used by the
Fedora 2 application.

Added a new endpoint, instead of modifying the "exists" endpoint
because handling multiple pairs of allowed parameters significantly
complicates the validation, and there is no easy way to express
the constraints in the OpenAPI specification.

Updated the "docs/umd-handle-open-api-v1.yml" OpenAPI specification for
the new endpoint.

https://issues.umd.edu/browse/LIBITD-1917