LIBITD-1908. Removed placeholder text from SAML config in env_example
Removed "lorem ipsum" placeholder text from the "SAML_SP_PRIVATE_KEY"
and "SAML_SP_CERTIFICATE" settings in the "env_example" file, to make it
completely obvious that they are simply placeholders and not actual
keys/certificates.

Removing the placeholder text also makes it easier to update a copied
".env" file when using "vi" on a server because there are fewer lines to
delete.

https://issues.umd.edu/browse/LIBITD-1908